Update set/flashcard/keyword database functions to overwrite Flashcard(sSet) structs
