+++ Sources/Model/Flashcard.swift
@@ -5,7 +5,7 @@
 
 import Foundation
 
-struct Flashcard: Identifiable, Codable {
+struct Flashcard: Identifiable {
 
     let id: Int64
     var front = ""
@@ -15,10 +15,11 @@ struct Flashcard: Identifiable, Codable {
     var tags: [String]?
     // swiftlint:enable discouraged_optional_collection
 
-    init(id: Int64, front: String = "", back: String = "") {
+    init(id: Int64, front: String = "", back: String = "", difficulty: Int64 = 0) {
         self.id = id
         self.front = front
         self.back = back
+        self.gameData.difficulty = Int(difficulty)
     }
 
     enum Side: String, Codable, CaseIterable, Identifiable {
